Changed: Migration to Java 21 (OFBIZ-13306)
Migration to Java 21

Thanks: Jacques for the warnings work and Nicolas.

# Conflicts:
#	framework/base/src/test/java/org/apache/ofbiz/base/util/collections/FlexibleMapAccessorTests.java
#	framework/base/src/test/java/org/apache/ofbiz/base/util/string/FlexibleStringExpanderTests.java

+++ framework/entity/src/main/java/org/apache/ofbiz/entity/util/EntityDataLoader.java
@@ -225,23 +225,23 @@ public class EntityDataLoader {
         return getUrlByComponentList(helperName, components, readerNames);
     }
 
-    public static int loadData(URL dataUrl, String helperName, Delegator delegator, List<Object> errorMessages) throws GenericEntityException {
+    public static long loadData(URL dataUrl, String helperName, Delegator delegator, List<Object> errorMessages) throws GenericEntityException {
         return loadData(dataUrl, helperName, delegator, errorMessages, -1);
     }
 
-    public static int loadData(URL dataUrl, String helperName, Delegator delegator, List<Object> errorMessages, int txTimeout)
+    public static long loadData(URL dataUrl, String helperName, Delegator delegator, List<Object> errorMessages, int txTimeout)
             throws GenericEntityException {
         return loadData(dataUrl, helperName, delegator, errorMessages, txTimeout, false, false, false);
     }
 
-    public static int loadData(URL dataUrl, String helperName, Delegator delegator, List<Object> errorMessages, int txTimeout, boolean dummyFks,
+    public static long loadData(URL dataUrl, String helperName, Delegator delegator, List<Object> errorMessages, int txTimeout, boolean dummyFks,
                                boolean maintainTxs, boolean tryInsert) throws GenericEntityException {
         return loadData(dataUrl, helperName, delegator, errorMessages, txTimeout, false, false, false, true);
     }
 
-    public static int loadData(URL dataUrl, String helperName, Delegator delegator, List<Object> errorMessages, int txTimeout, boolean dummyFks,
+    public static long loadData(URL dataUrl, String helperName, Delegator delegator, List<Object> errorMessages, int txTimeout, boolean dummyFks,
                                boolean maintainTxs, boolean tryInsert, boolean continueOnFail) throws GenericEntityException {
-        int rowsChanged = 0;
+        Long rowsChanged = (long) 0;
 
         if (dataUrl == null) {
             String errMsg = "Cannot load data, dataUrl was null";
